2nd_project/pyproject.toml
[tool.poetry]
name = "2nd-project"
version = "0.1.0"
description = ""
authors = ["Seou0912 <myuni0912@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
